+++ test/test_caster.py
@@ -20,10 +20,7 @@ def current_path():
 
 @pytest.fixture(scope="function")
 def modes():
-    return [
-        ("kg", "json"),
-        # ("ibes", "csv")
-    ]
+    return [("kg", "json"), ("ibes", "csv")]
 
 
 def cast(modes, schema_obj, current_path, level, reset, n_cores=1):